+++ pkg/compactor/bucket_compactor_e2e_test.go
@@ -206,7 +206,7 @@ func TestGroupCompactE2E(t *testing.T) {
 
 		ignoreDeletionMarkFilter := block.NewIgnoreDeletionMarkFilter(logger, objstore.WithNoopInstr(bkt), 48*time.Hour, fetcherConcurrency)
 		duplicateBlocksFilter := NewShardAwareDeduplicateFilter()
-		noCompactMarkerFilter := NewGatherNoCompactionMarkFilter(logger, objstore.WithNoopInstr(bkt), 2)
+		noCompactMarkerFilter := NewNoCompactionMarkFilter(logger, objstore.WithNoopInstr(bkt), 2, false)
 		metaFetcher, err := block.NewMetaFetcher(nil, 32, objstore.WithNoopInstr(bkt), "", nil, []block.MetadataFilter{
 			ignoreDeletionMarkFilter,
 			duplicateBlocksFilter,